Move EVAOperation based Tasks to new appropriate packages
CHanges:
- BudgetResources & DigLocal tasks & associated MetaTasks moved to structure/task package
- Maintain & loading of Vehicle tasks moved to vehcile/task package
- MetaTaskUtil updated with new MetaTask locations
- Callers from Mission classes updated
