+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/mission/AbstractVehicleMission.java
@@ -37,7 +37,6 @@ import com.mars_sim.core.person.ai.task.LoadVehicleGarage;
 import com.mars_sim.core.person.ai.task.LoadingController;
 import com.mars_sim.core.person.ai.task.OperateVehicle;
 import com.mars_sim.core.person.ai.task.Sleep;
-import com.mars_sim.core.person.ai.task.meta.LoadVehicleMeta;
 import com.mars_sim.core.person.ai.task.util.Task;
 import com.mars_sim.core.person.ai.task.util.TaskJob;
 import com.mars_sim.core.person.ai.task.util.TaskPhase;
@@ -55,6 +54,7 @@ import com.mars_sim.core.vehicle.StatusType;
 import com.mars_sim.core.vehicle.Vehicle;
 import com.mars_sim.core.vehicle.VehicleController;
 import com.mars_sim.core.vehicle.VehicleType;
+import com.mars_sim.core.vehicle.task.LoadVehicleMeta;
 import com.mars_sim.mapdata.location.Coordinates;
 import com.mars_sim.tools.Msg;
 import com.mars_sim.tools.util.RandomUtil;
