+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/mission/ConstructionMission.java
@@ -27,7 +27,6 @@ import com.mars_sim.core.mission.Construction;
 import com.mars_sim.core.person.Person;
 import com.mars_sim.core.person.ai.SkillType;
 import com.mars_sim.core.person.ai.task.ConstructBuilding;
-import com.mars_sim.core.person.ai.task.DigLocalRegolith;
 import com.mars_sim.core.person.ai.task.util.Task;
 import com.mars_sim.core.person.ai.task.util.Worker;
 import com.mars_sim.core.resource.ItemResourceUtil;
@@ -45,6 +44,7 @@ import com.mars_sim.core.structure.construction.ConstructionStage;
 import com.mars_sim.core.structure.construction.ConstructionStageInfo;
 import com.mars_sim.core.structure.construction.ConstructionValues;
 import com.mars_sim.core.structure.construction.ConstructionVehicleType;
+import com.mars_sim.core.structure.task.DigLocalRegolith;
 import com.mars_sim.core.vehicle.Crewable;
 import com.mars_sim.core.vehicle.GroundVehicle;
 import com.mars_sim.core.vehicle.LightUtilityVehicle;
