+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/mission/EmergencySupply.java
@@ -24,11 +24,9 @@ import com.mars_sim.core.malfunction.MalfunctionFactory;
 import com.mars_sim.core.malfunction.Malfunctionable;
 import com.mars_sim.core.person.Person;
 import com.mars_sim.core.person.ai.task.EVAOperation;
-import com.mars_sim.core.person.ai.task.UnloadVehicleEVA;
 import com.mars_sim.core.person.ai.task.UnloadVehicleGarage;
 import com.mars_sim.core.person.ai.task.Walk;
 import com.mars_sim.core.person.ai.task.WalkingSteps;
-import com.mars_sim.core.person.ai.task.meta.LoadVehicleMeta;
 import com.mars_sim.core.person.ai.task.util.Task;
 import com.mars_sim.core.person.ai.task.util.TaskJob;
 import com.mars_sim.core.person.ai.task.util.Worker;
@@ -39,6 +37,8 @@ import com.mars_sim.core.structure.building.Building;
 import com.mars_sim.core.time.MarsTime;
 import com.mars_sim.core.vehicle.Rover;
 import com.mars_sim.core.vehicle.Vehicle;
+import com.mars_sim.core.vehicle.task.LoadVehicleMeta;
+import com.mars_sim.core.vehicle.task.UnloadVehicleEVA;
 import com.mars_sim.mapdata.location.Coordinates;
 import com.mars_sim.mapdata.location.LocalPosition;
 import com.mars_sim.tools.util.RandomUtil;
