+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/mission/RoverMission.java
@@ -32,10 +32,8 @@ import com.mars_sim.core.person.ai.task.DriveGroundVehicle;
 import com.mars_sim.core.person.ai.task.EVAOperation;
 import com.mars_sim.core.person.ai.task.OperateVehicle;
 import com.mars_sim.core.person.ai.task.RequestMedicalTreatment;
-import com.mars_sim.core.person.ai.task.UnloadVehicleEVA;
 import com.mars_sim.core.person.ai.task.Walk;
 import com.mars_sim.core.person.ai.task.WalkingSteps;
-import com.mars_sim.core.person.ai.task.meta.UnloadVehicleMeta;
 import com.mars_sim.core.person.ai.task.util.Task;
 import com.mars_sim.core.person.ai.task.util.TaskJob;
 import com.mars_sim.core.person.ai.task.util.TaskPhase;
@@ -48,6 +46,8 @@ import com.mars_sim.core.vehicle.Rover;
 import com.mars_sim.core.vehicle.StatusType;
 import com.mars_sim.core.vehicle.Vehicle;
 import com.mars_sim.core.vehicle.VehicleType;
+import com.mars_sim.core.vehicle.task.UnloadVehicleEVA;
+import com.mars_sim.core.vehicle.task.UnloadVehicleMeta;
 import com.mars_sim.mapdata.location.LocalPosition;
 import com.mars_sim.tools.Msg;
 import com.mars_sim.tools.util.RandomUtil;
