+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/task/util/MetaTaskUtil.java
@@ -17,22 +17,17 @@ import java.util.stream.Collectors;
 import com.mars_sim.core.Simulation;
 import com.mars_sim.core.activities.GroupActivityMetaTask;
 import com.mars_sim.core.person.ai.task.meta.AnalyzeMapDataMeta;
-import com.mars_sim.core.person.ai.task.meta.BudgetResourcesMeta;
 import com.mars_sim.core.person.ai.task.meta.ConnectOnlineMeta;
 import com.mars_sim.core.person.ai.task.meta.ConsolidateContainersMeta;
 import com.mars_sim.core.person.ai.task.meta.ConstructBuildingMeta;
 import com.mars_sim.core.person.ai.task.meta.ConverseMeta;
 import com.mars_sim.core.person.ai.task.meta.CookMealMeta;
 import com.mars_sim.core.person.ai.task.meta.DelegateWorkMeta;
-import com.mars_sim.core.person.ai.task.meta.DigLocalIceMeta;
-import com.mars_sim.core.person.ai.task.meta.DigLocalRegolithMeta;
 import com.mars_sim.core.person.ai.task.meta.DoInventoryMeta;
 import com.mars_sim.core.person.ai.task.meta.EatDrinkMeta;
 import com.mars_sim.core.person.ai.task.meta.ExamineBodyMeta;
 import com.mars_sim.core.person.ai.task.meta.ListenToMusicMeta;
-import com.mars_sim.core.person.ai.task.meta.LoadVehicleMeta;
 import com.mars_sim.core.person.ai.task.meta.MaintainBuildingMeta;
-import com.mars_sim.core.person.ai.task.meta.MaintainVehicleMeta;
 import com.mars_sim.core.person.ai.task.meta.ManufactureConstructionMaterialsMeta;
 import com.mars_sim.core.person.ai.task.meta.ManufactureGoodMeta;
 import com.mars_sim.core.person.ai.task.meta.MeetTogetherMeta;
@@ -64,7 +59,6 @@ import com.mars_sim.core.person.ai.task.meta.TendGreenhouseMeta;
 import com.mars_sim.core.person.ai.task.meta.ToggleFuelPowerSourceMeta;
 import com.mars_sim.core.person.ai.task.meta.ToggleResourceProcessMeta;
 import com.mars_sim.core.person.ai.task.meta.TreatMedicalPatientMeta;
-import com.mars_sim.core.person.ai.task.meta.UnloadVehicleMeta;
 import com.mars_sim.core.person.ai.task.meta.WorkoutMeta;
 import com.mars_sim.core.person.ai.task.meta.WriteReportMeta;
 import com.mars_sim.core.person.ai.task.meta.YogaMeta;
@@ -81,6 +75,12 @@ import com.mars_sim.core.science.task.PerformMathematicalModelingMeta;
 import com.mars_sim.core.science.task.ProposeScientificStudyMeta;
 import com.mars_sim.core.science.task.RespondToStudyInvitationMeta;
 import com.mars_sim.core.science.task.StudyFieldSamplesMeta;
+import com.mars_sim.core.structure.task.BudgetResourcesMeta;
+import com.mars_sim.core.structure.task.DigLocalIceMeta;
+import com.mars_sim.core.structure.task.DigLocalRegolithMeta;
+import com.mars_sim.core.structure.task.MaintainVehicleMeta;
+import com.mars_sim.core.vehicle.task.LoadVehicleMeta;
+import com.mars_sim.core.vehicle.task.UnloadVehicleMeta;
 
 /**
  * A utility task for getting the list of meta tasks.
