+++ mars-sim-core/src/main/java/com/mars_sim/core/structure/task/DigLocalRegolith.java
@@ -1,0 +1,65 @@
+/*
+ * Mars Simulation Project
+ * DigLocalRegolith.java
+ * @date 2023-08-22
+ * @author Scott Davis
+ */
+
+package com.mars_sim.core.structure.task;
+
+import com.mars_sim.core.equipment.EquipmentType;
+import com.mars_sim.core.person.Person;
+import com.mars_sim.core.person.ai.task.util.TaskPhase;
+import com.mars_sim.core.resource.ResourceUtil;
+import com.mars_sim.tools.Msg;
+
+/**
+ * This class is a task for digging and collecting
+ * regolith right outside in the vicinity of a settlement.
+ */
+public class DigLocalRegolith
+extends DigLocal {
+
+	/** default serial id. */
+	private static final long serialVersionUID = 1L;
+
+	/** Simple Task name */
+	public static final String SIMPLE_NAME = DigLocalRegolith.class.getSimpleName();
+	
+	/** Task name */
+    public static final String NAME = Msg.getString(
+            "Task.description.digLocalRegolith"); //$NON-NLS-1$
+
+    /** Task phases. */
+    private static final TaskPhase COLLECT_REGOLITH = new TaskPhase(Msg.getString(
+            "Task.phase.collectRegolith")); //$NON-NLS-1$
+
+	/**
+	 * Constructor 1.
+	 * 
+	 * @param person the person performing the task.
+	 */
+	public DigLocalRegolith(Person person) {
+        // Use EVAOperation constructor.
+        super(NAME, COLLECT_REGOLITH, ResourceUtil.regolithID, 
+        	  EquipmentType.WHEELBARROW, person, 150);
+        if (!isDone()) {
+        	setCollectionRate(person.getAssociatedSettlement().getRegolithCollectionRate());
+        }
+    }
+	
+	/**
+	 * Constructor 2.
+	 * 
+	 * @param person the person performing the task.
+	 * @param duration
+	 */
+	public DigLocalRegolith(Person person, int duration) {
+        // Use EVAOperation constructor.
+        super(NAME, COLLECT_REGOLITH, ResourceUtil.regolithID, 
+        	  EquipmentType.WHEELBARROW, person, duration);
+        if (!isDone()) {
+        	setCollectionRate(person.getAssociatedSettlement().getRegolithCollectionRate());
+        }
+    }
+}
